+++ wallet-service/src/main/java/com/jack/walletservice/service/WalletService.java
@@ -4,17 +4,19 @@ import com.jack.common.dto.response.WalletBalanceDto;
 import com.jack.common.dto.response.WalletCreateMessageDto;
 import com.jack.walletservice.entity.Wallet;
 
+import java.math.BigDecimal;
+
 public interface WalletService {
 
     void createWallet(WalletCreateMessageDto message);
 
     Wallet getWalletByUserId(Long userId);
 
-    void updateWallet(Long userId, Double usdAmount, Double btcAmount);
+    void updateWallet(Long userId, BigDecimal usdAmount, BigDecimal btcAmount);
 
     void updateWalletBalance(WalletBalanceDto WalletBalanceDto);
 
-    void debitWallet(Long userId, Double amount);
+    void debitWallet(Long userId, BigDecimal amount);
 
     WalletBalanceDto getWalletBalance(Long userId);
